Added and fixed drops
Added ammo drops, therefore the player can
increase the amount of ammo in their weapons. Had to get rid of the event system for this as it made zero sense to use it for this purpose and it was causing issues. Moved about a few folders for better organization - still a nightmare though