Added puzzle, added new map and "fixed sword"
Added the combination puzzle where the correct levers will need to be pressed in a certain order to open the door. This is present in level 2, the new level.
The sword is really weird in in level 2, as in it does not want to do damage unless you aim for the feet?? bruh.